[518585] Provide a transformation manager view for CPS
This change provides the initial version of the view.
Target platform is updated to Oxygen and Java 8 is required
to allow using the new platform APIs in the following components:
 * generator.tests
 * xform.m2m.tests
 * xform.m2t.tests
 * application.ui

There is a known issue related to resetting incremental transformations:
incremental transformations doesn't reset the target model
therefore duplicates may appear on repeated transformation executions.

Change-Id: I493bcbaed5534aaf99ead8e3383f8afb2e8b5b8a
Signed-off-by: Marton Elekes <marci543@gmail.com>