[519597] Add debugger support to CPS Application
This commit adds support to run VIATRA CPS transformations with debugger
support.
The CPS Transformations view is restructured to provide a way to start a
transformation with or without debugging enabled.

Known issue: attaching transformation debugger and changing the model
from the same Eclipse instance causes the application to freeze.

Change-Id: Ie4c8a7228fcb97ba37ac3590152c3d08cd83233a
Signed-off-by: Marton Elekes <marci543@gmail.com>